% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyLayout.R
\name{applyLayout}
\alias{applyLayout}
\title{Apply a layout to the current network}
\usage{
applyLayout(layout, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{layout}{(char) name of the layut, run commandHelp('layout') to see available list of layouts.
Feel free to include parameters along with the layout name, space delimited. See example.}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Apply a layout to the current network
}
\section{Example}{
 applyLayout('force-directed defaultSpringCoefficient=.000004 defaultSpringLength=100')
}

