% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyLayoutSelected.R
\name{applyLayoutSelected}
\alias{applyLayoutSelected}
\title{Apply layout to just the selected nodes (Needs to be one of the layours available through cytoscape commands)
There are tunable parameters for attributes-layout layout}
\usage{
applyLayoutSelected(base.url, selected.layout = "degree-circle",
  max.width = 500, node.attribute = "__mclCluster")
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{selected.layout}{layout to be applied - default ot degree circle}

\item{max.width}{parameter used to attribtues circle layout that determines how many circles on x-axis - default to 500}

\item{node.attribute}{parameter used to attribtues circle layout that determines which attribute to use for circles}
}
\value{
server response
}
\description{
Apply layout to just the selected nodes (Needs to be one of the layours available through cytoscape commands)
There are tunable parameters for attributes-layout layout
}
