% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyStyle.R
\name{applyStyle}
\alias{applyStyle}
\title{Apply a style to the current network}
\usage{
applyStyle(style.name, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{style.name}{(char) name of the style}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Apply a style to the current network
}
\section{Example}{
 applyStyle("myStyle")
}

