% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundleEdges.R
\name{bundleEdges}
\alias{bundleEdges}
\title{Bundle edges}
\usage{
bundleEdges(network.suid = "current", base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network.suid}{suid of the network to apply edge bundling; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Bundle edges
}
\examples{
bundleEdges()
bundleEdges(network.suid="3570")
}
