% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCytoscapeVersion.R
\name{checkCytoscapeVersion}
\alias{checkCytoscapeVersion}
\title{Check the version of Cytoscape}
\usage{
checkCytoscapeVersion(base.url = "http://localhost:1234/v1")
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
Cytoscape version
}
\description{
Check the version of Cytoscape
}
