% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearSelection.R
\name{clearSelection}
\alias{clearSelection}
\title{Deselect all nodes and edges}
\usage{
clearSelection(network = "current", base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Deselect all nodes and edges
}
