% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command2query.R
\name{command2query}
\alias{command2query}
\title{Command string to CyREST query URL}
\usage{
command2query(cmd.string, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{cmd.string}{(char) command}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
cyrest url
}
\description{
Converts a command string to a CyREST query url.
}
\section{Example}{
 command2query('layout get preferred')
}

