% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandHelp.R
\name{commandHelp}
\alias{commandHelp}
\title{Command Help}
\usage{
commandHelp(cmd.string = "help", base.url = "http://localhost:1234/v1")
}
\arguments{
\item{cmd.string}{(char) command}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
List of available commands or args
}
\description{
Using the same syntax as Cytoscape's Command Line Dialog, 
this function returns a list of available commands or args.
}
\details{
Works with or without 'help' command prefix.
}
\examples{
commandHelp()  
commandHelp('node')
commandHelp('node get attribute')
}
