% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandRun.R
\name{commandRun}
\alias{commandRun}
\title{Command Run}
\usage{
commandRun(cmd.string, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{cmd.string}{(char) command}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
List object
}
\description{
Using the same syntax as Cytoscape's Command Line Dialog,
this function converts a command string into a CyREST URL, executes a GET
request, and parses the HTML result content into an R list object.
}
\section{Example}{
 commandRun('layout get preferred')
}

