% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEnrichmentmapGsea.R
\name{createEnrichmentmapGsea}
\alias{createEnrichmentmapGsea}
\title{Create Enrichment map with gsea results from edb directory}
\usage{
createEnrichmentmapGsea(base.url, name, pvalue_thresh, qvalue_thresh,
  similarity_thresh, similarity_metric, gmt_gsea_file, gsea_ranks_file,
  gsea_results_filename, exp_file)
}
\arguments{
\item{base.url}{- cyrest base url for communicating with cytoscape}

\item{name}{- name to call the new enrichment map network}

\item{pvalue_thresh}{- pvalue threhold}

\item{qvalue_thresh}{- FDR threshold}

\item{similarity_thresh}{- similarity threshold, between 0 and 1}

\item{similarity_metric}{- jaccard, overlap or combined}

\item{gmt_gsea_file}{- path to gmt file}

\item{gsea_results_filename}{- path to gsea edb file}

\item{exp_file}{- path to expression file}
}
\value{
suid of created enrichment map
}
\description{
Create Enrichment map with gsea results from edb directory
}
