% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNetwork.R
\name{createNetwork}
\alias{createNetwork}
\title{Create a network}
\usage{
createNetwork(nodes, edges, netName = "MyNetwork",
  collName = "myNetworkCollection", portNum = 1234,
  base.url = "http://localhost:1234/v1", ...)
}
\arguments{
\item{nodes}{(data.frame) see nodeSet2JSON() for details}

\item{edges}{(data.frame) see edgeSet2JSON() for details}

\item{netName}{(char) network name}

\item{collName}{(char) network collection name}

\item{portNum}{(int) port number for cytoscape (Deprecated)}

\item{base.url}{cyrest base url for communicating with cytoscape}

\item{...}{params for nodeSet2JSON() and edgeSet2JSON()}
}
\value{
(int) network ID
}
\description{
Takes data frames for nodes and edges, as well as naming parameters to
generate the JSON data format required by the "networks" POST operation via CyREST.
Returns the network.suid and applies the perferred layout set in Cytoscape preferences.
}
\seealso{
nodeSet2JSON edgeSet2JSON
}
