% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNetworkFromIgraph.R
\name{createNetworkFromIgraph}
\alias{createNetworkFromIgraph}
\title{Create a Cytoscape network from an igraph network}
\usage{
createNetworkFromIgraph(igraph, network.name = "MyNetwork",
  collection.name = "myNetworkCollection",
  base.url = "http://localhost:1234/v1", ...)
}
\arguments{
\item{igraph}{(igraph) igraph network object}

\item{network.name}{(char) network name}

\item{collection.name}{(char) network collection name}

\item{base.url}{cyrest base url for communicating with cytoscape}

\item{...}{params for nodeSet2JSON() and edgeSet2JSON(); see createNetwork}
}
\value{
(int) network SUID
}
\description{
Takes an igraph network and generates data frames for nodes and edges to
send to the generic createNetwork function.
Returns the network.suid and applies the perferred layout set in Cytoscape preferences.
}
\details{
Vertices and edges from the igraph network will be translated into nodes and edges
in Cytoscape. Associated attributes will also be passed to Cytoscape as node and edge table columns.
}
\section{Example}{
 createNetworkFromIgraph(g)
}

\seealso{
createNetwork
}
