% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyMoveSelectedNodes.R
\name{cyMoveSelectedNodes}
\alias{cyMoveSelectedNodes}
\title{Move all selected nodes x and y position by x.offset and y.offset}
\usage{
cyMoveSelectedNodes(base.url, network.suid, x.offset = 0, y.offset = 0,
  network.viewid)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{network.suid}{suid of the network that you want to get the view for}

\item{x.offset}{numeric that you want to add to every selected node x position - default is zero}

\item{y.offset}{numeric that you want to add to every selected node y position - default is zero}

\item{network.viewid}{suid of the network view that you want to get the info for}
}
\value{
server response
}
\description{
Move all selected nodes x and y position by x.offset and y.offset
}
