% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cySelectNodes.R
\name{cySelectNodes}
\alias{cySelectNodes}
\title{Select given nodes. Given a table that has a column called 'suids' method selects the nodes with supplied suids in}
\usage{
cySelectNodes(base.url, nodes, network.suid)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{nodes}{a matrix with nodes that you want to select.  The table needs to have a column with SUIDs}

\item{network.suid}{suid of the network that you want to get the view for}
}
\value{
server response
}
\description{
Select given nodes. Given a table that has a column called 'suids' method selects the nodes with supplied suids in
}
