% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyUnselectAllNodes.R
\name{cyUnselectAllNodes}
\alias{cyUnselectAllNodes}
\title{Unselect all nodes}
\usage{
cyUnselectAllNodes(base.url, network.suid)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{network.suid}{suid of the network that you want to get the view for}
}
\value{
server response
}
\description{
Unselect all nodes
}
