% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deselectAllNodes.R
\name{deselectAllNodes}
\alias{deselectAllNodes}
\title{Deselect all nodes}
\usage{
deselectAllNodes(network = "current", base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Deselect all nodes
}
