% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeSet2JSON.R
\name{edgeSet2JSON}
\alias{edgeSet2JSON}
\title{Convert edges to JSON format needed for CyRest network creation}
\usage{
edgeSet2JSON(edge_set, source.id.list = "source", target.id.list = "target",
  interaction.type.list = "interaction", ...)
}
\arguments{
\item{edge_set}{(data.frame) Rows contain pairwise interactions.}

\item{source.id.list}{(char) override default list name for source node ids}

\item{target.id.list}{(char) override default list name for target node ids}

\item{interaction.type.list}{(char) override default list name for interaction types
Columns contain edge attributes}
}
\description{
Convert edges to JSON format needed for CyRest network creation
}
\details{
Function takes a table of interactions. By default, edge_set
should contain columns: source, target and interaction. Additional columns
are loaded as edge attributes. The 'interaction' list can contain a single 
value to apply to all rows; and if excluded altogether, the interaction type
wiil be set to "interacts with".
}
