% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportImage.R
\name{exportImage}
\alias{exportImage}
\title{Exports the current network view as an image}
\usage{
exportImage(filename, type, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{filename}{(char) name of the image file to save}

\item{type}{(char) type of image to export, e.g., JPEG, PDF, PNG, PostScript, SVG (case sensitive)}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Exports the current network view as an image
}
\details{
The image is cropped per the current view in Cytoscape.
}
\section{Example}{
 exportImage('myNetwork','PDF')
}

