% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportNetwork.R
\name{exportNetwork}
\alias{exportNetwork}
\title{Exports the current network as a data file}
\usage{
exportNetwork(filename, type, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{filename}{(char) name of the network file to save}

\item{type}{(char) type of data file to export, e.g., CX, CYJS, GraphML, NNF, SIF, XGMML (case sensitive)}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Exports the current network as a data file
}
\section{Example}{
 exportNetwork('myNetwork','XGMML')
}

