% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportStyle.R
\name{exportStyle}
\alias{exportStyle}
\title{Exports the current visual style as a data file}
\usage{
exportStyle(filename, type, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{filename}{(char) name of the style file to save}

\item{type}{(char) type of data file to export, e.g., XML, JSON (case sensitive)}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Exports the current visual style as a data file
}
\section{Example}{
 exportStyle('myStyle','JSON')
}

