% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkName.R
\name{getNetworkName}
\alias{getNetworkName}
\title{Get the name of a network}
\usage{
getNetworkName(network.suid = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network.suid}{SUID of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
network viewid
}
\description{
Get the name of a network
}
\section{Examples}{
 getNetworkName() \cr
getNetworkName(1111)
}

