% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkSuid.R
\name{getNetworkSuid}
\alias{getNetworkSuid}
\title{Get the network suid}
\usage{
getNetworkSuid(network.name = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network.name}{name of the network that you want the suid; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
network viewid
}
\description{
Get the network suid
}
