% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeTable.R
\name{getNodeTable}
\alias{getNodeTable}
\title{Get node table}
\usage{
getNodeTable(network = "current", base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network}{name or suid of the network that you want the node table; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
matrix nodetable
}
\description{
Get node information from cytoscape,get the whole node table
}
