% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeviewId.R
\name{getNodeviewId}
\alias{getNodeviewId}
\title{get the network view suid}
\usage{
getNodeviewId(base.url, network.suid)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{network.suid}{suid of the network that you want the network view id}
}
\value{
network viewid
}
\description{
get the network view suid
}
