% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStyleNames.R
\name{getStyleNames}
\alias{getStyleNames}
\title{Get names of all visual styles}
\usage{
getStyleNames(base.url = "http://localhost:1234/v1")
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
network viewid
}
\description{
Get names of all visual styles
}
