% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableColumns.R
\name{getTableColumns}
\alias{getTableColumns}
\title{Get table column values}
\usage{
getTableColumns(table, columns, namespace = "default", network = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{table}{name of table, e.g., node, edge, network}

\item{columns}{names of columns to retrieve values from as list object or comma-separated list}

\item{namespace}{namespace of table; default is "default"}

\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
data frame of column values
}
\description{
Get table column values
}
\details{
The 'name' column is always retrieved along with specified columns. The 'name' values are used
as row names in the returned data frame. Note that this function fails on columns with missing values.
}
\section{Example}{
 getTableColumns('node','score')
}

