% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVisualProperties.R
\name{getVisualProperties}
\alias{getVisualProperties}
\title{Get the names and parameters of available Visual Properties}
\usage{
getVisualProperties(vis.prop = "all", base.url = "http://localhost:1234/v1")
}
\arguments{
\item{vis.prop}{name of visual property to query; default is "all"}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
visual properties
}
\description{
Get the names and parameters of available Visual Properties
}
