% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listNetworks.R
\name{listNetworks}
\alias{listNetworks}
\title{List the names of all networks in current session}
\usage{
listNetworks(base.url = "http://localhost:1234/v1")
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
list of network names
}
\description{
List the names of all networks in current session
}
