% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listStyles.R
\name{listStyles}
\alias{listStyles}
\title{List the names of all visual styles}
\usage{
listStyles(base.url = "http://localhost:1234/v1")
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
network viewid
}
\description{
List the names of all visual styles
}
