% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTableColumns.R
\name{listTableColumns}
\alias{listTableColumns}
\title{List table column names}
\usage{
listTableColumns(table, namespace = "default", network.name = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{table}{name of table, e.g., node, edge, network}

\item{namespace}{namespace of table, e.g., default}

\item{network.name}{name of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
list of column names
}
\description{
List table column names
}
\examples{
listTableColumns('node')
listTableColumns('edge')
listTableColumns('network')
}
