% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVisualProperties.R
\name{listVisualProperties}
\alias{listVisualProperties}
\title{List the names and parameters of available Visual Properties}
\usage{
listVisualProperties(vis.prop = "all",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{vis.prop}{name of visual property to query (case dependent); default is "all"}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
visual properties
}
\description{
List the names and parameters of available Visual Properties
}
