% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTableData.R
\name{loadTableData}
\alias{loadTableData}
\title{Loads data into Cytoscape tables keyed by row}
\usage{
loadTableData(data, data.key.column = "row.names", table = "node",
  table.key.column = "name", network = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{data}{(data.frame) each row is a node and columns contain node attributes}

\item{data.key.column}{(char) name of data.frame column to use as key; default is "row.names"}

\item{table}{(char) name of Cytoscape table to load data into, e.g., node, edge or network; default is "node"}

\item{table.key.column}{(char) name of Cytoscape table column to use as key; default is "name"}

\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\description{
This function loads data into Cytoscape node/edge/network tables provided a
common key, e.g., name. Data.frame column names will be used to set Cytoscape table column
names.
}
\details{
Numeric data columns will be stored as Doubles in Cytoscape tables. Character or mixed data
columns will be stored as Strings. Existing columns with the same names will be overwritten.
}
