% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEMapfromGMT.R
\name{makeEMapfromGMT}
\alias{makeEMapfromGMT}
\title{Make enrichment map from GMT}
\usage{
makeEMapfromGMT(gmtFile, simCutoff = 0.05, coefType = "JACCARD",
  portNum = 1234, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{gmtFile}{(char) path to GMT file}

\item{simCutoff}{(numeric from 0-1) similarity cutoff for node pruning}

\item{coefType}{(one of JACCARD or COMBINED) type of similarity measure}

\item{portNum}{(int) port number for cytoscape (Deprecated)}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
(int) network ID of newly created network
}
\description{
Make enrichment map from GMT
}
