% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_NodeFillDiscrete.R
\name{map_NodeFillDiscrete}
\alias{map_NodeFillDiscrete}
\title{creates node fill JSON to map node attribute to}
\usage{
map_NodeFillDiscrete(attrName, attrVals, cols)
}
\arguments{
\item{attrName}{(char) name of node attribute for discrete fill}

\item{attrVals}{(char) possible values}

\item{cols}{(char) corresponding colour mappings (e.g. "#3300CC")}
}
\value{
(list) ready to convert for JSON style mapping
}
\description{
creates node fill JSON to map node attribute to
}

