% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapEdgeWidthPassthrough.R
\name{mapEdgeWidthPassthrough}
\alias{mapEdgeWidthPassthrough}
\title{Creates edge width JSON for passthrough mapping of edge attribute}
\usage{
mapEdgeWidthPassthrough(attrName)
}
\arguments{
\item{attrName}{(char) name of edge attribute for passthrough mapping to edge width}
}
\value{
(list) ready to convert for JSON style mapping
}
\description{
Creates edge width JSON for passthrough mapping of edge attribute
}
