% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNodeBorderColorDiscrete.R
\name{mapNodeBorderColorDiscrete}
\alias{mapNodeBorderColorDiscrete}
\title{Creates node border color JSON for discrete mapping of node attribute}
\usage{
mapNodeBorderColorDiscrete(attrName, attrVals, cols)
}
\arguments{
\item{attrName}{(char) name of node attribute for discrete border color}

\item{attrVals}{(char) possible values}

\item{cols}{(char) corresponding color mappings (e.g. "#3300CC")}
}
\value{
(list) ready to convert for JSON style mapping
}
\description{
Creates node border color JSON for discrete mapping of node attribute
}
