% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNodeFillDiscrete.R
\name{mapNodeFillDiscrete}
\alias{mapNodeFillDiscrete}
\title{Creates node fill JSON for discrete mapping of node attribute}
\usage{
mapNodeFillDiscrete(attrName, attrVals, cols)
}
\arguments{
\item{attrName}{(char) name of node attribute for discrete fill}

\item{attrVals}{(char) possible values}

\item{cols}{(char) corresponding colour mappings (e.g. "#3300CC")}
}
\value{
(list) ready to convert for JSON style mapping
}
\description{
Creates node fill JSON for discrete mapping of node attribute
}
