% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNodeLabelPassthrough.R
\name{mapNodeLabelPassthrough}
\alias{mapNodeLabelPassthrough}
\title{Creates node label JSON for passthrough mapping of node attribute}
\usage{
mapNodeLabelPassthrough(attrName)
}
\arguments{
\item{attrName}{(char) name of node attribute for passthrough mapping to node label}
}
\value{
(list) ready to convert for JSON style mapping
}
\description{
Creates node label JSON for passthrough mapping of node attribute
}
