% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNodeSizeContinuous.R
\name{mapNodeSizeContinuous}
\alias{mapNodeSizeContinuous}
\title{Creates node size JSON for continuous mapping of node attribute}
\usage{
mapNodeSizeContinuous(attrName, attrVals, cols)
}
\arguments{
\item{attrName}{(char) name of node attribute for continuous size}

\item{attrVals}{(char) possible values}
}
\value{
(list) ready to convert for JSON style mapping
}
\description{
Creates node size JSON for continuous mapping of node attribute
}
