% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapStyleDefaults.R
\name{mapStyleDefaults}
\alias{mapStyleDefaults}
\title{Map style defaults}
\usage{
mapStyleDefaults(attrList)
}
\arguments{
\item{paramList}{(list) keys are attributes, values are values}
}
\value{
(list) data structure to generate JSON for default style mappings
}
\description{
Map style defaults
}
\details{
The full list of possible attributes will come soon.
For now, these are the available attributes:
NODE_TRANSPARENCY, NODE_BORDER_WIDTH, NODE_FILL_COLOR, 
* NODE_SHAPE: "diamond","ellipse","hexagon","octagon","parallelogram","rectangle","triangle","V",
* NODE_LABEL
EDGE_TRANSPARENCY, EDGE_WIDTH
NETWORK_BACKGROUND_PAINT
}
\examples{
mapStyleDefaults(list(NODE_FILL_COLOR="#FF0000",NODE_SIZE=12))
}
