% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapTargetArrowShapeDiscrete.R
\name{mapTargetArrowShapeDiscrete}
\alias{mapTargetArrowShapeDiscrete}
\title{Creates target arrow shape JSON for discrete mapping of edge attribute}
\usage{
mapTargetArrowShapeDiscrete(attrName, attrVals, shapes)
}
\arguments{
\item{attrName}{(char) name of edge attribute for discrete mapping to target arrow shape}

\item{attrVals}{(char) possible values}

\item{shapes}{(char) corresponding shape mappings (e.g. "None, Arrow, T, Circle, Square, Diamond")}
}
\value{
(list) ready to convert for JSON style mapping
}
\description{
Creates target arrow shape JSON for discrete mapping of edge attribute
}
