% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveSelectedNodes.R
\name{moveSelectedNodes}
\alias{moveSelectedNodes}
\title{Move selected nodes}
\usage{
moveSelectedNodes(x.offset = 0, y.offset = 0, network = "current",
  network.viewid = "current", base.url = "http://localhost:1234/v1")
}
\arguments{
\item{x.offset}{numeric that you want to add to every selected node x position - default is zero}

\item{y.offset}{numeric that you want to add to every selected node y position - default is zero}

\item{network}{name or suid of the network; default is "current" network}

\item{network.viewid}{suid of the network view; default is "current" network view}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Move all selected nodes x and y position by x.offset and y.offset.
}
