% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSet2JSON.R
\name{nodeSet2JSON}
\alias{nodeSet2JSON}
\title{Creates a table of nodes to CyREST JSON}
\usage{
nodeSet2JSON(node.set, node.id.list = "id", ...)
}
\arguments{
\item{node.set}{(data.frame) each row is a node and columns contain node
attributes}

\item{node.id.list}{(char) override default list name for node ids}
}
\description{
Creates a table of nodes to CyREST JSON
}
\examples{
mynodes <- data.frame(name=c("A","B","C","D"), 
		group=c("1","1","2","2"),
	stringsAsFactors=FALSE)
}
