% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCySwagger.R
\name{openCySwagger}
\alias{openCySwagger}
\title{Open CySwagger docs in browser}
\usage{
openCySwagger(domain = "cyrest", base.url = "http://localhost:1234/v1")
}
\arguments{
\item{domain}{(char) documentation domain or scope}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
Web page
}
\description{
Opens swagger docs in default browser for a live
instance of CyREST or CyREST-supported operations.
}
\section{Examples}{
 openCySwagger() \cr
openCySwagger('commands')
}

