% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2cytoscape-deprecated.R
\name{r2cytoscape-deprecated}
\alias{r2cytoscape-deprecated}
\alias{cyMoveSelectedNodes}
\alias{cySelectNodes}
\alias{getNodeviewId}
\alias{getNodeviewTable}
\alias{map_NodeFillDiscrete}
\alias{map_styleDefaults}
\alias{cyUnselectAllNodes}
\title{Deprecated}
\usage{
See newer versions of deprecated functions linked below
}
\description{
These functions are provided for compatibility with older version of
the r2cytoscape package.  They may eventually be completely
removed.
}
\section{Deprecation notices}{

\tabular{rl}{
  \code{cyMoveSelectedNodes} \tab has been replaced by \code{\link{moveSelectedNodes}}\cr
  \code{cySelectNodes} \tab has been replaced by \code{\link{selectNodes}}\cr
  \code{cyUnselectAllNodes} \tab has been replaced by \code{\link{unselectAllNodes}}\cr
  \code{getNodeviewId} \tab has been replaced by \code{\link{getNetworkViewId}}\cr
  \code{getNodeviewTable} \tab has been replaced by \code{\link{getNodeViewTable}}\cr
  \code{map_NodeFillDiscrete} \tab has been replaced by \code{\link{mapNodeFillDiscrete}}\cr
  \code{map_styleDefaults} \tab has been replaced by \code{\link{mapStyleDefaults}}\cr
}
}

