% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameNetwork.R
\name{renameNetwork}
\alias{renameNetwork}
\title{Rename network}
\usage{
renameNetwork(new.name, network.suid = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{new.name}{new name of the network}

\item{network.suid}{suid of the network that you want to change name; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Rename network
}
