% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAutoAnnotate.R
\name{runAutoAnnotate}
\alias{runAutoAnnotate}
\title{Run autoannotate analysis on network}
\usage{
runAutoAnnotate(base.url, cluster.alg = "MCL", max.words = 3,
  network.name = "none")
}
\arguments{
\item{base.url}{- cyrest base url for communicating with cytoscape}

\item{cluster.alg}{- cluster algorithm to use - default set to MCL
(options include - AFFINITY_PROPAGATION|CLUSTER_FIZZIFIER|GLAY|CONNECTED_COMPONENTS|MCL|SCPS)}

\item{max.words}{- max words to be used in each annotation - default set to 3.}

\item{network.name}{- name of network to perform post analysis on.}
}
\value{
server response
}
\description{
Run autoannotate analysis on network
}
