% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPostAnalysis.R
\name{runPostAnalysis}
\alias{runPostAnalysis}
\title{Run post analysis on network}
\usage{
runPostAnalysis(base.url, pa.name, pa.gmtFile, test = "MANN_WHIT_LESS",
  network.name = "none", threshold)
}
\arguments{
\item{base.url}{- cyrest base url for communicating with cytoscape}

\item{pa.name}{- name to call post analysis dataset}

\item{pa.gmtFile}{- path to post analysis gmt file}

\item{test}{- post analysis test you want to use - default MANN_WHIT_LESS
(options - NO_FILTER|HYPERGEOM|MANN_WHIT_TWO_SIDED|MANN_WHIT_GREATER|MANN_WHIT_LESS|NUMBER|PERCENT|SPECIFIC)}

\item{network.name}{- name of network to perform post analysis on.}

\item{threshold}{- post analysis edge threshold}
}
\value{
server response
}
\description{
Run post analysis on network
}
