% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSession.R
\name{saveSession}
\alias{saveSession}
\title{Saves the current Cytoscape session}
\usage{
saveSession(filename, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{filename}{(char) name of the session file to save}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Saves the current Cytoscape session
}
\details{
Saves session as a CYS file.
}
\section{Example}{
 saveSession('myFirstSession')
}

