% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFirstNeighbors.R
\name{selectFirstNeighbors}
\alias{selectFirstNeighbors}
\title{Select first neighbor nodes}
\usage{
selectFirstNeighbors(direction = "any", network = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{direction}{direction of connections to neighbors to follow, e.g., incoming, outgoing, undirected, or any (default)}

\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
list of names of selected nodes, including original selection
}
\description{
Select nodes directly connected to currently selected nodes. Can
specify connection directionality using the direction param.
}
