% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectNodes.R
\name{selectNodes}
\alias{selectNodes}
\title{Select nodes}
\usage{
selectNodes(nodes, network.suid, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{nodes}{a matrix with nodes that you want to select.  The table needs to have a column with SUIDs}

\item{network.suid}{suid of the network that you want to get the view for}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Select given nodes. Given a table that has a column called 'suids' method selects the nodes with supplied suids in
}
