% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCurrentNetwork.R
\name{setCurrentNetwork}
\alias{setCurrentNetwork}
\title{Set current network}
\usage{
setCurrentNetwork(network, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network}{name or suid of the network that you want set as current}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Selects the given network as "current"
}
