% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNodeSizeLocked.R
\name{setNodeSizeLocked}
\alias{setNodeSizeLocked}
\title{Set whether node width and height are locked}
\usage{
setNodeSizeLocked(locked, style, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{locked}{(boolean) TRUE enables a single NODE_SIZE property; FALSE enables separate NODE_WIDTH and NODE_HEIGHT properties}

\item{style}{name of visual style to apply dependency}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
