% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateStyleDependencies.R
\name{setStyleDependencies}
\alias{setStyleDependencies}
\title{Updates the values of dependencies in a style}
\usage{
setStyleDependencies(style.name, dependencies,
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{style.name}{(char) name for style}

\item{dependencies}{(list) a list of style dependencies, see list below. Note: 
each dependency is set by a boolean, TRUE or FALSE (T or F)}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Updates style dependencies, overriding any prior settings.
}
\section{List of Dependencies}{

arrowColorMatchesEdge
nodeCustomGraphicsSizeSync
nodeSizeLocked
}

\examples{
setStyleDependencies("myStyle",list(nodeSizeLocked=TRUE))
}
