% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateStyleMapping.R
\name{setStyleMapping}
\alias{setStyleMapping}
\title{Adds or updates a visual property mapping in a style}
\usage{
setStyleMapping(style.name, mapping, base.url = "http://localhost:1234/v1")
}
\arguments{
\item{style.name}{(char) name for style}

\item{mapping}{a single visual property mapping, see mapVisualProperty}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Creates a visual property mapping if it doesn't already exist in the style. Otherwise,
updates the visual property mapping, overriding any prior mapping.
}
\details{
Requires visual property mappings to be previously created, see mapVisualProperty.
}
\examples{
setStyleMapping('myStyle',mapVisualProperty('node label','name','p'))
}
\seealso{
mapVisualProperty
}
