% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unselectAllNodes.R
\name{unselectAllNodes}
\alias{unselectAllNodes}
\title{Unselect all nodes}
\usage{
unselectAllNodes(network.suid = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network.suid}{suid of the network that you want to get the view for; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Unselect all nodes
}
