% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateStyleDefaults.R
\name{updateStyleDefaults}
\alias{updateStyleDefaults}
\title{Updates the default values of visual properties in a style}
\usage{
updateStyleDefaults(style.name, defaults,
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{style.name}{(char) name for style}

\item{defaults}{(list) a list of visual property default settings, see mapVisualProperty}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Updates visual property defaults, overriding any prior settings. See mapVisualProperty for
the list of visual properties.
}
\section{Example}{
 updateStyleDefaults('myStyle',list('node fill color'='#0000FF','node size'=50))
}

\seealso{
mapVisualProperty
}
