% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandHelp.R
\name{commandHelp}
\alias{commandHelp}
\title{Command Help}
\usage{
commandHelp(cmd.string = "help", base.url = "http://localhost:1234/v1")
}
\arguments{
\item{cmd.string}{(char) command}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
List of available commands or args
}
\description{
Using the same syntax as Cytoscape's Command Line Dialog,
this function returns a list of available commands or args.
}
\details{
Works with or without 'help' command prefix. Note that if you ask about a command that doesn't
have any arguments, this function will run the command!
}
\section{Examples}{
 commandHelp() \cr
commandHelp('node') \cr
commandHelp('node get attribute')
}

