% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNetwork.R
\name{createNetwork}
\alias{createNetwork}
\title{create a network}
\usage{
createNetwork(nodes, edges, netName = "network", collName = "collection",
  portNum = 1234, ...)
}
\arguments{
\item{nodes}{(data.frame) see nodeSet2JSON() for details}

\item{edges}{(data.frame) see edgeSet2JSON() for details}

\item{netName}{(char) network name}

\item{collName}{(char) network collection name}

\item{portNum}{(int) port number for cytoscape}

\item{...}{params for nodeSet2JSON() and edgeSet2JSON()}
}
\value{
(int) network ID
}
\description{
create a network
}

