% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStyle.R
\name{createStyle}
\alias{createStyle}
\title{create a style from components}
\usage{
createStyle(styleName, defaults, mappings, portNum = 1234,
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{styleName}{(char) name for style}

\item{defaults}{(list) key-value pairs for default mappings. See
mapStyleDefaults() for details on formatting of this variable}

\item{mappings}{(list) attribute mappings. Each entry should be 
a specific attribute mapping (e.g. discrete fill color)}

\item{portNum}{(int) port number for cytoscape (Deprecated)}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
URL of style and JSON. Side effect of POSTing Style.
}
\description{
create a style from components
}
\details{
requires attribute mappings to be previously created
}
