% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStyle.R
\name{createStyle}
\alias{createStyle}
\title{Create a style from components}
\usage{
createStyle(style.name, defaults, mappings, portNum = 1234,
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{style.name}{(char) name for style}

\item{defaults}{(list) key-value pairs for default mappings. See
mapStyleDefaults() for details on formatting of this variable}

\item{mappings}{(list) visual property mappings, see mapVisualProperty}

\item{portNum}{(int) port number for cytoscape (Deprecated)}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Creates a style from defaults and predefined mappings.
}
\details{
Requires attribute mappings to be previously created, see mapVisualProperty.
}
\seealso{
mapVisualProperty
}
