% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStyle.R
\name{createStyle}
\alias{createStyle}
\title{Create a style from components}
\usage{
createStyle(style.name, defaults, mappings,
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{style.name}{(char) name for style}

\item{defaults}{(list) key-value pairs for default mappings.}

\item{mappings}{(list) visual property mappings, see mapVisualProperty}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
None
}
\description{
Creates a style from defaults and predefined mappings.
}
\details{
Requires attribute mappings to be previously created, see mapVisualProperty.
}
\section{Example construction}{

\preformatted{
style.name = "myStyle"
defaults <- list(NODE_SHAPE="diamond",
                 NODE_SIZE=30,
                 EDGE_TRANSPARENCY=120,
                 NODE_LABEL_POSITION="W,E,c,0.00,0.00")
nodeLabels <- mapVisualProperty('node label','id','p')
nodeFills <- mapVisualProperty('node fill color','group','d',c("A","B"), c("#FF9900","#66AAAA"))
arrowShapes <- mapVisualProperty('Edge Target Arrow Shape','interaction','d',c("activates","inhibits","interacts"),c("Arrow","T","None"))
edgeWidth <- mapVisualProperty('edge width','weight','p')

createStyle(style.name, defaults, list(nodeLabels,nodeFills,arrowShapes,edgeWidth))
}
}

\seealso{
mapVisualProperty
}
