% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeSet2JSON.R
\name{edgeSet2JSON}
\alias{edgeSet2JSON}
\title{Convert edges to JSON format needed for CyRest network creation}
\usage{
edgeSet2JSON(interactions, interA = "AliasA", interB = "AliasB",
  inter_type = "method", ...)
}
\arguments{
\item{interactions}{(data.frame) Rows contain pairwise interactions.
Columns contain edge attributes}
}
\description{
Convert edges to JSON format needed for CyRest network creation
}
\details{
Function takes a table of interactions.  Interaction table needs to
contain columns: AliasA, AliasB, and pmids
}
