% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkSuid.R
\name{getNetworkSuid}
\alias{getNetworkSuid}
\title{Get the SUID of a network}
\usage{
getNetworkSuid(network.name = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network.name}{name of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
network viewid
}
\description{
Get the SUID of a network
}
\section{Examples}{
 getNetworkSuid() \cr
getNetworkSuid("myNetwork")
}

