% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkViewId.R
\name{getNetworkViewId}
\alias{getNetworkViewId}
\title{Get the network view suid}
\usage{
getNetworkViewId(base.url = "http://localhost:1234/v1", network.suid)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{network.suid}{suid of the network that you want the network view id}
}
\value{
network viewid
}
\description{
Get the network view suid
}
