% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkViewId.R
\name{getNetworkViewId}
\alias{getNetworkViewId}
\title{Get the network view suid}
\usage{
getNetworkViewId(network = "current", base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
network viewid
}
\description{
Get the network view suid
}
