% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeTable.R
\name{getNodeTable}
\alias{getNodeTable}
\title{Get node information from cytoscape,get the whole node table}
\usage{
getNodeTable(base.url, network.suid)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{network.suid}{suid of the network that you want the node table}
}
\value{
matrix nodetable
}
\description{
Get node information from cytoscape,get the whole node table
}
