% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeTable.R
\name{getNodeTable}
\alias{getNodeTable}
\title{Get node table}
\usage{
getNodeTable(network = "current", base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network}{name or suid of the network that you want the node table; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
matrix node table
}
\description{
Retrieve the entire node table as a matrix object
}
\details{
If there are missing values in the Node Table, then a warning message about the rbind function
and the "number of columns of result is not a multiple of vector length". This indicates that values were
shifted (and repeated) to fill in missing columns. This could result in corrupted data for certain nodes.
In these cases, consider using getTableColumns instead.
}
\seealso{
getTableColumns
}
