% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeViewTable.R
\name{getNodeViewTable}
\alias{getNodeViewTable}
\title{Get Node View Table}
\usage{
getNodeViewTable(base.url = "http://localhost:1234/v1", network.suid,
  network.viewid)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{network.suid}{suid of the network that you want to get the view for}

\item{network.viewid}{suid of the network view that you want to get the info for}
}
\value{
matrix nodeview parameters x-position, y-position, selected
}
\description{
Get node view information from cytoscape, get the node view table - x and y positions
}
