% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeviewTable.R
\name{getNodeViewTable}
\alias{getNodeViewTable}
\title{Get node view table}
\usage{
getNodeViewTable(network = "current", network.viewid = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{network}{name or suid of the network; default is "current" network}

\item{network.viewid}{suid of the network view that you want to get the info for; default is "current" network view}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
matrix nodeview parameters x-position, y-position, selected
}
\description{
Get node view information from cytoscape, get the node view table - x and y positions
}
