% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeviewTable.R
\name{getNodeviewTable}
\alias{getNodeviewTable}
\title{Get view node information from cytoscape, get the node view table - x and y positions}
\usage{
getNodeviewTable(base.url, network.suid, network.viewid)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{network.suid}{suid of the network that you want to get the view for}

\item{network.viewid}{suid of the network view that you want to get the info for}
}
\value{
matrix nodeview parameters x-position, y-position, selected
}
\description{
Get view node information from cytoscape, get the node view table - x and y positions
}
