% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTableColumns.R
\name{listTableColumns}
\alias{listTableColumns}
\title{List names of all columns in a table}
\usage{
listTableColumns(table, namespace = "default", network = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{table}{name of table, e.g., node, edge, network}

\item{namespace}{namespace of table, e.g., default}

\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
list of column names
}
\description{
List names of all columns in a table
}
\section{Examples}{
 listTableColumns('node') \cr
listTableColumns('edge') \cr
listTableColumns('network')
}

