% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEMapfromGMT.R
\name{makeEMapfromGMT}
\alias{makeEMapfromGMT}
\title{make enrichment map from GMT}
\usage{
makeEMapfromGMT(gmtFile, simCutoff = 0.05, coefType = "JACCARD",
  portNum = 1234)
}
\arguments{
\item{gmtFile}{(char) path to GMT file}

\item{simCutoff}{(numeric from 0-1) similarity cutoff for node pruning}

\item{coefType}{(one of JACCARD or COMBINED) type of similarity measure}

\item{portNum}{(int) port number for Cytoscape}
}
\value{
(int) network ID of newly created network
}
\description{
make enrichment map from GMT
}

