% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSet2JSON.R
\name{nodeSet2JSON}
\alias{nodeSet2JSON}
\title{Creates a table of nodes to CyREST JSON}
\usage{
nodeSet2JSON(node_set, nodeID_column = "Alias", ...)
}
\arguments{
\item{node_set}{(data.frame) each row is a node and columns contain node
attributes}

\item{nodeID_column}{(char) column of table to set for id,name and 
shared_name}
}
\description{
Creates a table of nodes to CyREST JSON
}
\examples{
mynodes <- data.frame(Alias=c("A","B","C","D"), 
		GROUP=c("YES","YES","NO","NO"),
	stringsAsFactors=FALSE)
json_nodes <- convert_nodes_tojson(mynodes)
}
