% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameNetwork.R
\name{renameNetwork}
\alias{renameNetwork}
\title{Rename network}
\usage{
renameNetwork(new.name, network = "current",
  base.url = "http://localhost:1234/v1")
}
\arguments{
\item{new.name}{new name of the network}

\item{network}{name or suid of the network to rename; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Rename network
}
\section{Examples}{
 renameNetwork("newName") \cr
renameNetwork("newName","oldName")
}

