% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameNetwork.R
\name{renameNetwork}
\alias{renameNetwork}
\title{Rename network}
\usage{
renameNetwork(base.url, new.name, network.suid)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{new.name}{new name of the network}

\item{network.suid}{suid of the network that you want to change name}
}
\value{
server response
}
\description{
Rename network
}
