% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unselectAllNodes.R
\name{unselectAllNodes}
\alias{unselectAllNodes}
\title{Unselect all nodes}
\usage{
unselectAllNodes(base.url = "http://localhost:1234/v1", network.suid)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}

\item{network.suid}{suid of the network that you want to get the view for}
}
\value{
server response
}
\description{
Unselect all nodes
}
